feat: add jwt authentication

+++ src/main/java/com/lavertis/tasktrackerapi/controllers/UserController.java
@@ -3,19 +3,19 @@ package com.lavertis.tasktrackerapi.controllers;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.github.fge.jsonpatch.JsonPatch;
 import com.github.fge.jsonpatch.JsonPatchException;
-import com.lavertis.tasktrackerapi.dto.CreateUserRequest;
 import com.lavertis.tasktrackerapi.entities.User;
+import com.lavertis.tasktrackerapi.exceptions.BadRequestException;
 import com.lavertis.tasktrackerapi.exceptions.NotFoundException;
 import com.lavertis.tasktrackerapi.services.user_service.IUserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/users")
+@RequestMapping("/api/user")
 public class UserController {
 
     final IUserService userService;
@@ -36,12 +36,6 @@ public class UserController {
         return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
-    @PostMapping
-    public ResponseEntity<User> create(@Valid @RequestBody CreateUserRequest request) {
-        var user = userService.create(request);
-        return new ResponseEntity<>(user, HttpStatus.CREATED);
-    }
-
     @PatchMapping(value = "{id}", consumes = "application/json-patch+json")
     public ResponseEntity<User> updateById(@PathVariable long id, @RequestBody JsonPatch patch) throws JsonPatchException, JsonProcessingException, NotFoundException {
         var user = userService.updateById(id, patch);
@@ -49,7 +43,12 @@ public class UserController {
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<Void> deleteById(@PathVariable long id) {
+    public ResponseEntity<Void> deleteById(@PathVariable long id) throws NotFoundException, BadRequestException {
+        String username = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
+        User user = userService.findByUsername(username);
+        if (user.getId() != id)
+            throw new BadRequestException("Id of the request principal does not match passed parameter id");
+
         userService.deleteById(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
